fix status code of having 0 value
